refactor(#6) - Coupon 도메인 수정, 에러 코드 추가, 메서드 참조 식으로 변경

+++ src/main/java/com/flab/infrun/coupon/presentation/response/EnrolledCouponResponse.java
@@ -1,12 +1,13 @@
 package com.flab.infrun.coupon.presentation.response;
 
 import com.flab.infrun.coupon.application.result.EnrolledCouponResult;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record EnrolledCouponResponse(
     String ownerEmail,
     String discountType,
-    int discountValue,
+    BigDecimal discountValue,
     LocalDateTime expirationAt
 ) {
